Bug 861039 - Update ANGLE to r2042. - r=bjacob

+++ gfx/angle/src/compiler/MapLongVariableNames.cpp
@@ -9,7 +9,7 @@
 
 namespace {
 
-TString mapLongName(int id, const TString& name, bool isGlobal)
+TString mapLongName(size_t id, const TString& name, bool isGlobal)
 {
     ASSERT(name.size() > MAX_SHORTENED_IDENTIFIER_SIZE);
     TStringStream stream;
@@ -77,7 +77,7 @@ void LongNameMap::Insert(const char* originalName, const char* mappedName)
         originalName, mappedName));
 }
 
-int LongNameMap::Size() const
+size_t LongNameMap::Size() const
 {
     return mLongNameMap.size();
 }
@@ -122,7 +122,7 @@ TString MapLongVariableNames::mapGlobalLongName(const TString& name)
     const char* mappedName = mGlobalMap->Find(name.c_str());
     if (mappedName != NULL)
         return mappedName;
-    int id = mGlobalMap->Size();
+    size_t id = mGlobalMap->Size();
     TString rt = mapLongName(id, name, true);
     mGlobalMap->Insert(name.c_str(), rt.c_str());
     return rt;
